[SW-0] tests: Rename test_dma_all2all in gadui_dma.c
As there is a usage now of DEVICE_DONT_CARE in the gaudi_*.c files, the
corresponding binaries won't return immediately when being run on
another ASIC type, and each test will be skipped explicitly, and
test_dma_all2all among others.
"run_hlthunk_tests" counts the "SKIPPED" lines but ignores repetitions,
so the summary in this case will lack one of the skipped tests, and will
have an indication for a test with an unknown result.

To avoid this situation, simply rename the test in gadui_dma.c.

If in the future we change the handling in the gaudi_*.c files to not
using DEVICE_DONT_CARE, then this test can be renamed back to its
original name, and in spite of the duplicity.

Change-Id: I0cebd6fa6142b5b869cafe19145e907f0e35b934
Signed-off-by: Tomer Tayar <ttayar@habana.ai>
Signed-off-by: Oded Gabbay <ogabbay@habana.ai>